[RSDK-6687] Movement is stoppable/cancelable in the middle (#4)
Tried on my laptop, using the RC cards on app.viam.com. If you start a long move command and then click the Stop button, it stops. If you start a long move command and then navigate away from the Control page to the Config page, it stops after about a second.


* it's a motor, not a board

* don't wait to get the mutex before stopping the motor

* actually stop the motor in Stop()