Get the motor to go backwards when the RC card asks for it (#5)
The pre-existing code works great when you send a `GoFor` with a negative distance and positive speed. However, the RC card tells you to go backwards by sending a positive distance at a negative speed! So, support that version, too.

I made one of the tests move 0.1 revolutions instead of 0.01 revolutions, while investigating where the issue was, and I prefer this way because you can actually see it move, rather than twitch in place. but I can undo that change if you want.



* add todo: unused offset when resetting encoder

* abstract out named constant

* the speed should always be positive

* rephrase comment

* remove the TODO about the offset, it's fixed elsewhere