[RSDK-6727] Update readme (#11)
This has a couple other tweaks in it, but it's mostly the readme.

I'm also grouping the min/max acceleration together and the min/max deceleration together in the config, rather than min accel/decel and then max accel/decel. In hindsight, that's the way people are likely to think about these values.

In addition, the stopping deceleration rate should be the deceleration rate set by the user, and not an acceleration rate! For reasons I don't understand, certain motors can accelerate faster than they can decelerate, so keep these separate.



* Describe how to configure your network correctly for the device

* Add in descriptions of the min/max accel/decel options

* remove trailing whitespace

* Describe the extra accel/decel values you can set

* Explain how to skip having limits

* add TODO about how to implement SetPower. This is unrelated to the 'real' change

* min_rpm and max_rpm are not optional

* describe how to set unspecified parameters

* typo

* The stopping deceleration should always be the default deceleration, never the acceleration if it differs

* tweak: group min/max accelerations together, and min/max decelerations together

* clean up comment that discusses something removed in a previous PR

* clarify which IP address you set in your network settings

* link to the manual for the hardware, and describe how to use the DoCommand